Re-use fixtures between unit and integration tests
